Remove Round call from captureSnapshot
This doesn't remove the delta rounding, which is probably erroneous and
faulty. Need to test that more. In the meantime, this'll just assign to
the slot nearest when the snapshot was said to have been gathered
(provided by whence).

There's probably a better way of doing this. Plus it's not currently
doing anything to remain thread-safe, so there are obviously a bunch of
issues here.