Fix getBuffer/WriteTo not knowing about tempBuffers
Simple mistake, but resulted in tempBuffers trying to write to
themselves, which isn't really useful.

Fixed by adding getWriter, which will unwrap a *tempBuffer and return
a *bytes.Buffer for it. This way we can just check if two writers are
the same and not care all that much otherwise.

Additionally, this means that getBuffer(*tempBuffer) will result in
a new unowned *tempBuffer, which allows stacking them on top of one
another -- this makes it pretty easy to undo writes to them.