Add Dup methods to Tags and Fields
This replaces use of direct map[string]{string,Field} for tags and
fields, respectively, with the Tags and Fields types. As a result, it
means I don't need to write the usual duplicate-fields stuff as often
as I would, which is fine by me.

This changes the method signature for Fields() and Tags() on the
Measurement interface, so it is a breaking change.