Add RawPoint, export fixedInt/Bool/Float as RawInt/Bool/Float
RawString is a special case since fixedString is actually a pre-escaped
byte slice, so it'd be misleading to export that in its current state.
As such, RawString escapes on write. I could probably add a Compile()
method or something to RawString to return a fixedString Field, but
that's really splitting hairs over something I don't feel worth doing
just now.