Add missing ctx parameter to SendMeasurements
Keeps with the Proxy writer's mode of using contexts to handle proxy
cancellation and such. In this case, it's just passed through to
ctxhttp, so nothing special is going on.

If you pass a nil context, the function panics, since I can't decide
whether a default context should be created for you or if this is
a logical error in your code.