Send HTTP requests in the background so we're not blocking future sends
This is to prevent buffer build-ups because a prior send didn't fail
yet (but may soon time out or something). As a result, the actual HTTP
request will be done in a separate goroutine and the read buffer will
be let go of as soon as is reasonable (i.e., the buffer is copied for
the request body's use).