Break the outflux proxy ctor
This removes the Close method on Proxy and the general-purpose context
in use by the Proxy. You're now responsible for providing a context to
Proxy.Start if you want to signal cancellation or a deadline on the
thing. If you call Flush, you can pass a context there as well (useful
if you've opted not to allow auto-flushes of any kind).

This also adds some experimental auto-flushing around write sizes. By
default, it's disabled, to keep with the current expected behavior of
just flushing on an interval.

Change-Id: Ib00822d4dc1d06e54bbee7f2b17c78ec3f2c2f29