Replace dubb with bufferchain
dubb has some weird bugs, and although this won't alleviate all of the
bugs in the Proxy in general, it will result in slightly better
blocking behavior when writing measurements. It does, however, require
atomic writes to the bufferchain, which dagr should already accomplish
by virtue of it building up the whole line for a point before writing
it out.

bufferchain is a much simpler piece of code, though it's a little weird
conceptually (only a little). It's effectively just a buffer pool. When
a write is made to it, it pulls one of its buffers out and writes to
it. This ensures that a buffer is usually available and, if not, it
shouldn't be blocking for too long. Right now outflux proxies will
default to using 6 buffers or floor(1.7 x NumCPU) for the number of
buffers. This is likely pessimistic but should be fine for most cases.
More buffers increases the time flush() takes but reduces contention on
writes, so I don't really have good numbers right now aside from the
environments I'm normally using this in. This will necessarily use more
memory than before.

This removes the use of proxyWriter in favor of overriding the closer
on the lockbuffer returned by *bufferchain.Writer.

Change-Id: I450ab4c17361b295c20a518b7d384b87add7a1db