Implement request retries / limitations
This adds the ability to limit requests sent out to a specific number
of retries, with only a certain number of requests in-flight at a time,
and with a configurable backoff.

It also expands BadStatusError a bit for manual flushes, which probably
isn't really necessary but might be useful.

A BadStatusError is returned when Influx returns anything other than
204 now, as their docs specifically state that status 200 will return
an error message (I'm not entirely sure why an error isn't a 400).

Change-Id: I154ef956d7abbe0ba160a03e29b48bb39aba34cc