Update README, LICENSE.txt
Copies the LICENSE.txt to outflux so that it's clear they're both
licensed the same.

Rewrap the license and format the code to be longer vertically but
otherwise the same (i.e., doesn't require a scrollbar, ideally). This
means adjusting the examples a bit to shrink the output.

Contributing section updated to account for gerrit usage.

Also remove copy of license from README.

[ci skip]

Change-Id: I8d1b9dbd343677164d3331b9d84713c135a08a7f