Make excess-size flushes synchronous
This is to prevent a race to flush things that I'm hoping is avoidable,
since otherwise it would spawn too many goroutines too often. This can
slow down writes to the buffer and the InfluxDB server, but shouldn't
take too long.

This works mainly by taking a lock when an excess flush size occurs and
the flush size is known to be large enough to flush. Once the lock is
acquired, the buffer size is checked again. Then two scenarios are
possible:

- The buffer size is no longer in excess of the flush size, so the lock
  is released and the function returns. This means another goroutine
  acquired the lock first and did a swap and send. As a result, this is
  effectively the nop case.

- The buffer size is still in excess of the flush size, so pass the
  unlock function forward to swapAndSend. Once the swap occurs, the
  unlock function is called and the lock is released. At this point,
  all other goroutines waiting on the buffer should either see the
  buffer size decreased or the buffer rapidly grew and another
  goroutine initiates the swap.

It's pretty hacky, and this will need to be ported forward to the
protocol handling rewrite, but this allows outflux to be a little less
trigger-happy on requests.