Adjust flush locking, send timer
Fixes a bug with the send timer where outflux would lock itself up on
trying to drain the send timer. This has been modified to ensure that
it's impossible that we'll reset the timer before it's already been
drained, since flushes no longer reset it.

Unlocking of the flushlock is essentially guaranteed at this point, but
shouldn't ever be necessary due to current structure. Still, it makes
the code a little simpler to think about.