Break early on dispatch retries
This fixes a bug where the loop would wait the next subsequent backoff
time if one was given (i.e., the backoff function would control whether
it waited after the last failure). This is undesirable because it
blocks the dispatch queue longer than is needed.

This adds a mid-loop check for retries and breaks early if it failed
its final dispatch attempt.

Change-Id: Ia98540af126c406d6ac5896fd0bbbe3d22ec0111